Updated root info page layout
